Resolve issue with setting MIDI Channels on altered devices
There was an edge case where if a device that didn't need a MIDI channel was
swapped out for one that did (or vice versa) the Library would throw an Exception.

This was down to the fact that we weren't setting a MIDI Channel (or DeviceID)
until after changing the device (the try...catch not withstanding).

The setting to `None` on devices that no longer need a Channel/DeviceID is not
necessary, but is essentially house-cleaning of the `MIDIFixture` instance.